test route user/projet/contributeur ok

+++ src/api/views.py
@@ -5,9 +5,11 @@ from rest_framework.pagination import PageNumberPagination
 from rest_framework import status
 from rest_framework.permissions import IsAuthenticated
 from .models import User, Project, Contributor, Issue, Comment
-from .serializers import UserSerializer, ProjectSerializer, ContributorSerializer, IssueSerializer, CommentSerializer, ProjectDetailSerializer
+from .serializers import (
+    UserSerializer, ProjectSerializer, ContributorSerializer,
+    IssueSerializer, CommentSerializer, ProjectDetailSerializer
+    )
 from .permissions import ProjectPermission, ContributorPermission
-
 
 
 class UserViewSet(viewsets.ModelViewSet):
@@ -16,20 +18,28 @@ class UserViewSet(viewsets.ModelViewSet):
     permission_classes = [IsAuthenticated]
 
     # Droit a l'oubli
-    @action(detail=True, methods=['delete'], permission_classes=[IsAuthenticated])
+    @action(
+        detail=True,
+        methods=['delete'],
+        permission_classes=[IsAuthenticated]
+        )
     def delete_user(self, request, pk=None):
         user = self.get_object()
 
         # Supprimer les projets où l'utilisateur est l'auteur
         Project.objects.filter(author_id=user).delete()
-        
+
         # Supprimer les contributions de l'utilisateur
         Contributor.objects.filter(user_id=user).delete()
 
         # Supprimer l'utilisateur lui-même
         user.delete()
 
-        return Response({'status': 'user deleted'}, status=status.HTTP_204_NO_CONTENT)
+        return Response(
+            {'status': 'user deleted'},
+            status=status.HTTP_204_NO_CONTENT
+            )
+
 
 class ProjectViewSet(viewsets.ModelViewSet):
     queryset = Project.objects.all()
@@ -39,7 +49,11 @@ class ProjectViewSet(viewsets.ModelViewSet):
     def perform_create(self, serializer):
         serializer.save(author_id=self.request.user)
 
-    @action(detail=False, methods=['get'], permission_classes=[IsAuthenticated])
+    @action(
+        detail=False,
+        methods=['get'],
+        permission_classes=[IsAuthenticated]
+        )
     def project_summary(self, request):
         paginator = PageNumberPagination()
         paginator.page_size = 10
@@ -48,38 +62,69 @@ class ProjectViewSet(viewsets.ModelViewSet):
         serializer = ProjectDetailSerializer(result_page, many=True)
         return paginator.get_paginated_response(serializer.data)
 
+
 class ContributorViewSet(viewsets.ModelViewSet):
     queryset = Contributor.objects.all()
     serializer_class = ContributorSerializer
     permission_classes = [IsAuthenticated]
 
-    @action(detail=False, methods=['delete'], permission_classes=[IsAuthenticated])
+    def create(self, request, *args, **kwargs):
+        # Récupérer le project_id depuis les données de la requête
+        project_id = request.data.get('project_id')
+        project = Project.objects.get(pk=project_id)
+
+        # Vérifier que l'utilisateur est bien l'auteur du projet
+        if project.author_id != request.user:
+            return Response(
+                {'detail': "Vous n'êtes pas autorisé à ajouter des contributeurs à ce projet."},
+                status=status.HTTP_403_FORBIDDEN
+            )
+
+        # Si tout va bien, procéder à la création du contributeur
+        return super().create(request, *args, **kwargs)
+
+    @action(
+        detail=False,
+        methods=['delete'],
+        permission_classes=[IsAuthenticated, ProjectPermission]
+        )
     def remove_contributor(self, request):
         user_id = request.data.get('user_id')
         project_id = request.data.get('project_id')
-        try:
-            user = User.objects.get(pk=user_id)
-            project = Project.objects.get(pk=project_id)
-            contributor = Contributor.objects.get(project_id=project, user_id=user)
-            contributor.delete()
-            return Response({'status': 'contributor removed'}, status=status.HTTP_204_NO_CONTENT)
-        except User.DoesNotExist:
-            return Response({'error': 'user not found'}, status=status.HTTP_404_NOT_FOUND)
-        except Project.DoesNotExist:
-            return Response({'error': 'project not found'}, status=status.HTTP_404_NOT_FOUND)
-        except Contributor.DoesNotExist:
-            return Response({'error': 'contributor not found'}, status=status.HTTP_404_NOT_FOUND)
+
+        # Verif que utilisateur et projet existent
+        user = User.objects.get(pk=user_id)
+        project = Project.objects.get(pk=project_id)
+
+        # Vérifier que l'utilisateur est bien l'auteur du projet
+        if project.author_id != request.user:
+            return Response(
+                {'detail': "Vous n'êtes pas autorisé à supprimer des contributeurs de ce projet."},
+                status=status.HTTP_403_FORBIDDEN
+            )
+        
+        # verif que le contributeur existe
+        contributor = Contributor.objects.get(project_id=project, user_id=user)
+
+        #suppriem contributeur
+        contributor.delete()
+        return Response(
+            {'status': 'contributor removed'},
+            status=status.HTTP_204_NO_CONTENT
+            )
+
 
 class IssueViewSet(viewsets.ModelViewSet):
     queryset = Issue.objects.all()
     serializer_class = IssueSerializer
     permission_classes = [IsAuthenticated, ContributorPermission]
-    
+
 
 class CommentViewSet(viewsets.ModelViewSet):
     queryset = Comment.objects.all()
     serializer_class = CommentSerializer
     permission_classes = [IsAuthenticated, ContributorPermission]
+
 
 class ProjectListView(generics.ListAPIView):
     queryset = Project.objects.all()
